make base class
