+++ src/backend/TrafficCourtsApi/DisputeApi.Web/Features/TicketLookup/ITicketDisputeService.cs
@@ -11,19 +11,11 @@ namespace DisputeApi.Web.Features.TicketLookup
         public Task<TicketDispute> RetrieveTicketDisputeAsync(string ticketNumber, string time);
     }
 
-    public class TicketDisputeFromRsiService : ITicketDisputeService
+    public abstract class TicketDisputeService : ITicketDisputeService
     {
-        private readonly IRsiRestApi _rsiApi;
-        public TicketDisputeFromRsiService(IRsiRestApi rsiApi)
-        {
-            _rsiApi = rsiApi;
-        }
-
         public async Task<TicketDispute> RetrieveTicketDisputeAsync(string ticketNumber, string time)
         {
-            RawTicketSearchResponse rawResponse = await _rsiApi.GetTicket(
-                new GetTicketParams { TicketNumber = ticketNumber, PRN = "10006", IssuedTime = time.Replace(":", "") }
-            );
+            RawTicketSearchResponse rawResponse =  await GetTicket(ticketNumber, time);
             if (rawResponse == null || rawResponse.Items == null || rawResponse.Items.Count == 0) return null;
 
             foreach (Item item in rawResponse.Items)
@@ -32,35 +24,14 @@ namespace DisputeApi.Web.Features.TicketLookup
                 int lastSlash = item.SelectedInvoice.Reference.LastIndexOf('/');
                 if (lastSlash <= 0) continue;
                 string invoiceNumber = item.SelectedInvoice.Reference.Substring(lastSlash + 1);
-                Invoice invoice = await _rsiApi.GetInvoice(invoiceNumber);
+                Invoice invoice = await GetInvoice(invoiceNumber);
                 item.SelectedInvoice.Invoice = invoice;
             }
 
             return rawResponse.ConvertToTicketDispute();
         }
-    }
 
-    public class TicketDisputeFromFilesService : ITicketDisputeService
-    {
-        public async Task<TicketDispute> RetrieveTicketDisputeAsync(string ticketNumber, string time)
-        {
-            await using FileStream openStream = File.OpenRead($"Features/TicketLookup/ticket-{ticketNumber}.json");
-            RawTicketSearchResponse rawResponse =
-                await JsonSerializer.DeserializeAsync<RawTicketSearchResponse>(openStream);
-
-            if (rawResponse == null)
-            {
-                return null;
-            }
-
-            for (int i = 0; i < rawResponse.Items.Count; i++)
-            {
-                await using FileStream itemStream =
-                    File.OpenRead($"Features/TicketLookup/invoice-{ticketNumber}{i + 1}.json");
-                var invoice = await JsonSerializer.DeserializeAsync<Invoice>(itemStream);
-                rawResponse.Items[i].SelectedInvoice.Invoice = invoice;
-            }
-            return rawResponse.ConvertToTicketDispute();
-        }
+        protected abstract Task<RawTicketSearchResponse> GetTicket(string ticketNumber, string time);
+        protected abstract Task<Invoice> GetInvoice(string invoiceNumber);
     }
 }
